Don't use full screen loading message when loading replays.
The big flash of black is a bit jarring so instead black out the
folder list and show loading on the right hand side.
